Add tpsDecimalPlace, msptDecimalPlace configuration.

+++ src/main/java/df11zomgraves/ingameinfo/util/StringConvertUtils.java
@@ -40,7 +40,7 @@ public class StringConvertUtils {
 		}
 	}
 	
-	public static String getFloatDisplayFormat(float value, int decimalPlace) {
+	public static String getFloatDisplayFormat(double value, int decimalPlace) {
 		if (decimalPlace < 0)
 			decimalPlace = 0;
 		else if (decimalPlace > 6)
